requirements_api.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.4.0
autocommand==2.2.2
av==16.0.1
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
bitsandbytes==0.48.2
build==1.3.0
cachetools==6.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
chromadb==1.3.5
click==8.3.1
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.3
cut-cross-entropy==25.1.1
cycler==0.12.1
datasets==4.3.0
debugpy==1.8.16
decorator==5.2.1
diffusers==0.35.2
dill==0.4.0
distro==1.9.0
docstring-parser==0.17.0
durationpy==0.10
executing==2.2.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.7.0
gdown==5.2.0
google-auth==2.43.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-transfer==0.1.9
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
inflect==7.3.1
inquirerpy==0.3.4
ipykernel==6.30.1
ipython-pygments-lexers==1.1.1
ipython==9.5.0
ipywidgets==8.1.7
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jedi==0.19.2
jinja2==3.1.4
jsonschema-specifications==2025.9.1
jsonschema==4.25.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
keras==3.12.0
kiwisolver==1.4.9
kubernetes==34.1.0
libclang==18.1.1
markdown-it-py==4.0.0
markdown==3.10
markupsafe==2.1.5
matplotlib-inline==0.1.7
matplotlib==3.10.7
mdurl==0.1.2
ml-dtypes==0.5.4
mmh3==5.2.0
more-itertools==10.3.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python==4.11.0.86
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opt-einsum==3.4.0
optree==0.18.0
orjson==3.11.4
overrides==7.7.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.18.0
pexpect==4.9.0
pfzy==0.3.4
pillow==11.0.0
pip==25.2
platformdirs==4.4.0
posthog==5.4.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pyasn1-modules==0.4.2
pyasn1==0.6.1
pybase64==1.4.2
pydantic-core==2.41.5
pydantic==2.12.5
pygments==2.19.2
pyparsing==3.2.5
pypika==0.48.9
pyproject-hooks==1.2.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyyaml==6.0.2
pyzmq==27.1.0
qwen-vl-utils==0.0.14
referencing==0.37.0
regex==2025.11.3
requests-oauthlib==2.0.0
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
shtab==1.8.0
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
sympy==1.13.3
tenacity==9.1.2
tensorboard-data-server==0.7.2
tensorboard==2.20.0
tensorflow==2.20.0
termcolor==3.2.0
tf-keras==2.20.1
timm==1.0.22
tokenizers==0.22.1
tomli==2.0.1
torch==2.9.0
torchao==0.14.1
torchaudio==2.8.0+cu128
torchvision==0.24.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.2
triton==3.5.0
trl==0.24.0
typeguard==4.4.4
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tyro==0.9.35
tzdata==2025.2
ultralytics-thop==2.0.18
ultralytics==8.3.0
unsloth-zoo==2025.11.6
unsloth==2025.11.6
urllib3==2.3.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
watermark==2.5.0
wcwidth==0.2.13
websocket-client==1.9.0
websockets==15.0.1
werkzeug==3.1.4
wheel==0.45.1
widgetsnbextension==4.0.14
wrapt==2.0.1
xformers==0.0.33.post1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0